# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API страницы гардероба внутри Telegram. Пользователя сервер узнаёт по подписанным данным запуска Mini App, идентификатор из тела и параметров не берётся.",
        "title": "Daily Stylist Mini App API",
        "contact": {},
        "version": "0.7.0"
    },
    "basePath": "/",
    "paths": {
        "/api/cities": {
            "get": {
                "security": [
                    {
                        "initData": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Город и настройки"
                ],
                "summary": "Найти город по названию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Часть названия, не короче limits.min_city_query из /api/options",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miniapp.citiesResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/cities/at": {
            "get": {
                "security": [
                    {
                        "initData": []
                    }
                ],
                "description": "Координаты округляются до сотых, около километра: погоде точнее не нужно.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Город и настройки"
                ],
                "summary": "Определить город по координатам",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Широта",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Долгота",
                        "name": "longitude",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miniapp.cityBody"
                        }
                    },
                    "400": {
                        "description": "bad_request",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "404": {
                        "description": "place_not_found",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/city": {
            "put": {
                "security": [
                    {
                        "initData": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Город и настройки"
                ],
                "summary": "Сохранить город пользователя",
                "parameters": [
                    {
                        "description": "Город",
                        "name": "city",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miniapp.cityBody"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Город сохранён"
                    },
                    "400": {
                        "description": "bad_request",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "422": {
                        "description": "invalid_location",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/items": {
            "get": {
                "security": [
                    {
                        "initData": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Гардероб"
                ],
                "summary": "Все вещи пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miniapp.itemsResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "initData": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Гардероб"
                ],
                "summary": "Добавить вещь",
                "parameters": [
                    {
                        "description": "Вещь",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miniapp.itemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/miniapp.itemResponse"
                        }
                    },
                    "400": {
                        "description": "bad_request",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "409": {
                        "description": "wardrobe_full",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "422": {
                        "description": "invalid_item",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/items/delete": {
            "post": {
                "security": [
                    {
                        "initData": []
                    }
                ],
                "description": "Не DELETE с телом: тело у DELETE не определено стандартом, и прокси вправе его выбросить.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Гардероб"
                ],
                "summary": "Удалить вещи",
                "parameters": [
                    {
                        "description": "Идентификаторы вещей",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miniapp.deleteItemsRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Вещи удалены"
                    },
                    "400": {
                        "description": "bad_request",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/items/{id}": {
            "put": {
                "security": [
                    {
                        "initData": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Гардероб"
                ],
                "summary": "Изменить вещь",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор вещи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Вещь",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miniapp.itemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miniapp.itemResponse"
                        }
                    },
                    "400": {
                        "description": "bad_request",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "404": {
                        "description": "not_found",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    },
                    "422": {
                        "description": "invalid_item",
                        "schema": {
                            "$ref": "#/definitions/miniapp.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "miniapp.citiesResponse": {
            "type": "object",
            "properties": {
                "cities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/miniapp.cityBody"
                    }
                }
            }
        },
        "miniapp.cityBody": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "miniapp.deleteItemsRequest": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "miniapp.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "not_found"
                }
            }
        },
...
        "miniapp.itemsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/miniapp.itemResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "initData": {
            "description": "tma \u003cinitData\u003e - данные запуска Mini App, подписанные токеном бота",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}